Allow allocations to be exported as a CSV
This feature adds an additional action to the tools dropdown of session settings.
It allows the allocation to be downloaded as a CSV.

The CSV format is as close to the course budget format as I could quickly do,
all information is available and spreadsheet magic can fill in the gaps.